requirements.txt
# dependencies of risiti
jinja2
django-registration
psycopg2
pil
